+++ app/src/main/java/com/akg/akg_sales/view/dialog/ItemQuantityDialog.java
@@ -14,15 +14,15 @@ public class ItemQuantityDialog {
     DialogItemQtyBinding binding;
     public String itemName;
     public String itemUom;
-    public Integer quantity;
+    public Double quantity;
     public String btnLabel;
     private Context context;
     private Dialog dialog;
-    Consumer<Integer> callback;
+    Consumer<Double> callback;
 
     public ItemQuantityDialog(
             Context context, String itemName,String itemUom,
-            Integer quantity,String btnLabel,Consumer<Integer> callback){
+            Double quantity,String btnLabel,Consumer<Double> callback){
         this.context = context;
         this.itemName = itemName;
         this.itemUom = itemUom;
@@ -39,7 +39,7 @@ public class ItemQuantityDialog {
 
     public void onClickAddToCart(){
         try {
-            int qty = Integer.parseInt(binding.quantity.getText().toString());
+            double qty = Double.parseDouble(binding.quantity.getText().toString());
             if(qty<=0) throw new Exception("Invalid Quantity");
             this.callback.accept(qty);
             dialog.dismiss();
